Refactor transfer package structure
