+++ shared/src/androidMain/kotlin/at/asitplus/wallet/app/android/iso/transfer/capability/NfcInfo.android.kt
@@ -1,4 +1,4 @@
-package at.asitplus.wallet.app.common.iso.transfer.capability
+package at.asitplus.wallet.app.common.iso.transfer.method
 
 import android.content.BroadcastReceiver
 import android.content.Context
