+++ shared/src/commonMain/kotlin/at/asitplus/wallet/app/common/iso/transfer/method/DeviceEngagementMethods.kt
@@ -1,4 +1,4 @@
-package at.asitplus.wallet.app.common.iso.transfer.capability
+package at.asitplus.wallet.app.common.iso.transfer.method
 
 import androidx.compose.material.icons.Icons
 import androidx.compose.ui.graphics.vector.ImageVector
