+++ shared/src/commonMain/kotlin/at/asitplus/wallet/app/common/iso/transfer/method/openAppSettings.kt
@@ -1,4 +1,4 @@
-package at.asitplus.wallet.app.common.iso.transfer.capability
+package at.asitplus.wallet.app.common.iso.transfer.method
 
 /**
  * Opens app settings page so the user can manage permissions.
