+++ shared/src/commonMain/kotlin/at/asitplus/wallet/app/common/iso/transfer/state/PreconditionState.kt
@@ -1,4 +1,4 @@
-package at.asitplus.wallet.app.common.iso.transfer.capability
+package at.asitplus.wallet.app.common.iso.transfer.state
 
 enum class PreconditionState {
     NO_TRANSFER_METHOD_SELECTED,
