+++ shared/src/commonMain/kotlin/at/asitplus/wallet/app/common/iso/transfer/state/VerifierState.kt
@@ -1,4 +1,4 @@
-package at.asitplus.wallet.app.common.iso.transfer.capability
+package at.asitplus.wallet.app.common.iso.transfer.state
 
 sealed class VerifierState {
     object Init : VerifierState()
