+++ shared/src/commonMain/kotlin/ui/views/iso/holder/ShowQrCodeView.kt
@@ -33,12 +33,12 @@ import at.asitplus.valera.resources.heading_label_show_qr_code_screen
 import at.asitplus.valera.resources.info_text_qr_code_loading
 import at.asitplus.valera.resources.info_text_transfer_settings_loading
 import at.asitplus.wallet.app.common.iso.transfer.MdocHelper
-import at.asitplus.wallet.app.common.iso.transfer.capability.DeviceTransferMethodManager
-import at.asitplus.wallet.app.common.iso.transfer.capability.PreconditionState
-import at.asitplus.wallet.app.common.iso.transfer.capability.ShowQrCodeState
-import at.asitplus.wallet.app.common.iso.transfer.capability.TransferPrecondition
-import at.asitplus.wallet.app.common.iso.transfer.capability.rememberPlatformContext
-import at.asitplus.wallet.app.common.iso.transfer.capability.rememberTransferSettingsState
+import at.asitplus.wallet.app.common.iso.transfer.method.DeviceTransferMethodManager
+import at.asitplus.wallet.app.common.iso.transfer.method.rememberPlatformContext
+import at.asitplus.wallet.app.common.iso.transfer.state.PreconditionState
+import at.asitplus.wallet.app.common.iso.transfer.state.ShowQrCodeState
+import at.asitplus.wallet.app.common.iso.transfer.state.TransferPrecondition
+import at.asitplus.wallet.app.common.iso.transfer.state.rememberTransferSettingsState
 import io.github.aakira.napier.Napier
 import io.github.alexzhirkevich.qrose.options.QrBrush
 import io.github.alexzhirkevich.qrose.options.QrColors
