+++ shared/src/commonMain/kotlin/ui/views/iso/verifier/VerifierView.kt
@@ -10,12 +10,12 @@ import at.asitplus.valera.resources.Res
 import at.asitplus.valera.resources.info_text_check_response
 import at.asitplus.valera.resources.info_text_transfer_settings_loading
 import at.asitplus.valera.resources.info_text_waiting_for_response
-import at.asitplus.wallet.app.common.iso.transfer.capability.DeviceTransferMethodManager
-import at.asitplus.wallet.app.common.iso.transfer.capability.PreconditionState
-import at.asitplus.wallet.app.common.iso.transfer.capability.TransferPrecondition
-import at.asitplus.wallet.app.common.iso.transfer.capability.VerifierState
-import at.asitplus.wallet.app.common.iso.transfer.capability.rememberPlatformContext
-import at.asitplus.wallet.app.common.iso.transfer.capability.rememberTransferSettingsState
+import at.asitplus.wallet.app.common.iso.transfer.method.DeviceTransferMethodManager
+import at.asitplus.wallet.app.common.iso.transfer.method.rememberPlatformContext
+import at.asitplus.wallet.app.common.iso.transfer.state.PreconditionState
+import at.asitplus.wallet.app.common.iso.transfer.state.TransferPrecondition
+import at.asitplus.wallet.app.common.iso.transfer.state.VerifierState
+import at.asitplus.wallet.app.common.iso.transfer.state.rememberTransferSettingsState
 import io.github.aakira.napier.Napier
 import org.jetbrains.compose.resources.stringResource
 import org.multipaz.compose.permissions.rememberBluetoothPermissionState
