+++ shared/src/iosMain/kotlin/at/asitplus/wallet/app/common/iso/transfer/method/openAppSettings.ios.kt
@@ -1,4 +1,4 @@
-package at.asitplus.wallet.app.common.iso.transfer.capability
+package at.asitplus.wallet.app.common.iso.transfer.method
 
 import platform.Foundation.NSURL
 import platform.UIKit.UIApplication
